typecast
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@5182 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/bzfs.cxx
@@ -2675,7 +2675,7 @@ void zapFlag(int flagIndex)
 // Try to penalize winning casuality 
 static float rabbitRank (PlayerInfo& player) {
   if (clOptions->rabbitSelection == RandomRabbitSelection)
-    return bzfrand();
+    return (float)bzfrand();
   
   // otherwise do score-based ranking
   int sum = player.wins + player.losses;
